% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_est.R
\name{alpha_est}
\alias{alpha_est}
\title{Estimate \eqn{\alpha} from \eqn{\hat{\Omega}}}
\usage{
alpha_est(
  P_hat,
  A,
  gamma_hat,
  d,
  est_method = "avg",
  Omega_init = NULL,
  xtol = 1e-05
)
}
\arguments{
\item{P_hat}{(n-by-n-by-T array) Estimated time series of the transition probability matrix}

\item{A}{(n-by-n-by-T array) time series of the adjacency matrices}

\item{gamma_hat}{The estimated autoregressive parameter}

\item{d}{dimension of the latent position}

\item{est_method}{estimation method of \eqn{\Omega} matrix. "avg" (direct estimation) or "opt" (via minimum distance)}

\item{Omega_init}{= NUll by default: initial value}

\item{xtol}{= 1.0e-5 by default: numerical accuracy}
}
\value{
A list containing the following:
\describe{
\item{`alpha_est`}{(n-by-d matrix) Estimated latent positions}
\item{`Omega_hat`}{(n-by-n matrix) Estimated \eqn{\Omega} matrix}
}
}
\description{
Estimate \eqn{\alpha} from \eqn{\hat{\Omega}}
}
