% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_est.R
\name{avg_omega}
\alias{avg_omega}
\title{Estimation of \eqn{\hat{\Omega}} by averaging}
\usage{
avg_omega(P_hat, A, gamma_hat)
}
\arguments{
\item{P_hat}{(n-by-n-by-T array) Estimated time series of the transition probability matrix}

\item{A}{(n-by-n-by-T array) time series of the adjacency matrices}

\item{gamma_hat}{The estimated autoregressive parameter}
}
\value{
\eqn{\Omega} matrix
}
\description{
Estimation of \eqn{\hat{\Omega}} by averaging
}
