% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMLE.R
\name{CMLE_est}
\alias{CMLE_est}
\title{Estimate \eqn{\gamma} by CMLE}
\usage{
CMLE_est(A, gamma_init = NULL, xtol = 1e-05)
}
\arguments{
\item{A}{(n-by-n-by-T array) time series of the adjacency matrices}

\item{gamma_init}{initial value of gamma}
}
\value{
\eqn{\hat{gamma}}
}
\description{
Estimate \eqn{\gamma} by CMLE
}
