% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_est.R
\name{deriv_dist_ij}
\alias{deriv_dist_ij}
\title{Derivative of the entry-wise distance function}
\usage{
deriv_dist_ij(omega, p, a, gamma)
}
\arguments{
\item{omega}{scalar}

\item{p}{vector of length TT}

\item{a}{vector of length TT}

\item{gamma}{The autoregressive parameter}
}
\value{
distance_ij
}
\description{
Derivative of the entry-wise distance function
}
