% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMLE.R
\name{deriv_loglike_CMLE}
\alias{deriv_loglike_CMLE}
\title{Derivative of the Log likelihood function for the conditional
maximum likelihood estimation}
\usage{
deriv_loglike_CMLE(A, gamma)
}
\arguments{
\item{A}{(n-by-n-by-T array) time series of the adjacency matrices}

\item{gamma}{The autoregressive parameter}
}
\value{
derivative of the log-likelihood
}
\description{
Derivative of the Log likelihood function for the conditional
maximum likelihood estimation
}
