% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinnet.R
\docType{package}
\name{dinnet}
\alias{dinnet}
\title{dinnet: package for estimation of dynamic network models with fixed effects}
\description{
Include the estimation functions, data generation processes and the Monte
Carlo routine
}
