% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_est.R
\name{dist_ij}
\alias{dist_ij}
\title{entry-wise distance function}
\usage{
dist_ij(omega, p, a, gamma)
}
\arguments{
\item{omega}{scalar}

\item{p}{vector of length TT}

\item{a}{vector of length TT}

\item{gamma}{The autoregressive parameter}
}
\value{
distance_ij
}
\description{
entry-wise distance function
}
