% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdpg_est.R
\name{estimate_ASE}
\alias{estimate_ASE}
\title{Function for estimation of P with ASE period by period}
\usage{
estimate_ASE(A_t, d, sim = NULL, bfcheck = TRUE, l = 1e-04, u = 0.9999)
}
\arguments{
\item{A_t}{(n-by-n) Adjacency matrix}

\item{d}{dimension of the latent position}

\item{sim}{The random seed}

\item{bfcheck}{Check whether all entries of the estimated probabiltiy matrix are between 0 and 1}

\item{l}{Lower bound of the probability matrix, i.e. set all entries that not greater than 0 to `l`. 0.0001 by default}

\item{u}{Upper bound of the probability matrix, i.e. set all entries that not less than 1 to be `u`. 0.9999 by default.}
}
\value{
estimated P
}
\description{
Function for estimation of P with ASE period by period
}
\section{Warning}{
 random seeds changed
}

