% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdpg_est.R
\name{estimate_GRDPG}
\alias{estimate_GRDPG}
\title{Function for estimation of P with GRDPG}
\usage{
estimate_GRDPG(A, d, sim = NULL)
}
\arguments{
\item{A}{(n-by-n-by-T array) time series of the adjacency matrices}

\item{d}{dimension of the latent position}

\item{sim}{The random seed}
}
\value{
estimated P array for all time periods
}
\description{
Function for estimation of P with GRDPG
}
\section{Warning}{
 random seeds changed
}

