% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdpg_est.R
\name{estimateX_ASE}
\alias{estimateX_ASE}
\title{Function for estimation of X with ASE}
\usage{
estimateX_ASE(A, d, sim = NULL)
}
\arguments{
\item{d}{dimension of the latent position}

\item{sim}{The random seed}

\item{A_t}{(n-by-n) Adjacency matrix}
}
\value{
estimated X_hat
}
\description{
Function for estimation of X with ASE
}
\section{Warning}{
 random seeds changed
}

