% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Function to generate the network time series}
\usage{
generate_data(latent_positions, gamma, time_periods, sim)
}
\arguments{
\item{latent_positions}{(n-by-d Matrix) The alpha matrix}

\item{gamma}{(\eqn{|\gamma| < 1}) The autoregressive parameter}

\item{time_periods}{number of time periods T}

\item{sim}{The random seed}
}
\value{
A list containing the following:
\describe{
\item{`A0`}{(n-by-n matrix) Initial adjacency matrix \eqn{A_0}}
\item{`P0`}{(n-by-n matrix) Initial transition probability matrix \eqn{P_0}}
\item{`A`}{(n-by-n-by-T array) time series of the adjacency matrices}
\item{`P`}{(n-by-n-by-T array) time series of the transition probability matrix}
\item{`alpha`}{(n-by-n matrix) the product of latent positions \eqn{a_i^T a_j}}
}
}
\description{
The function generates the (stationary) time series of the networks
without covariates \eqn{x_{ij}}
}
\section{Warning}{
 random seeds changed
}

