% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMLE.R
\name{loglike_CMLE}
\alias{loglike_CMLE}
\title{Log likelihood function for the conditional maximum likelihood estimation}
\usage{
loglike_CMLE(A, gamma)
}
\arguments{
\item{A}{(n-by-n-by-T array) time series of the adjacency matrices}

\item{gamma}{The autoregressive parameter}
}
\value{
log-likelihood
}
\description{
Log likelihood function for the conditional maximum likelihood estimation
}
