% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_est.R
\name{min_dist_omega}
\alias{min_dist_omega}
\title{Minimize distance function to estimation \eqn{\hat{\Omega}}}
\usage{
min_dist_omega(P_hat, A, gamma_hat, Omega_init = NULL, xtol = 1e-05)
}
\arguments{
\item{P_hat}{(n-by-n-by-T array) Estimated time series of the transition probability matrix}

\item{A}{(n-by-n-by-T array) time series of the adjacency matrices}

\item{gamma_hat}{The estimated autoregressive parameter}

\item{Omega_init}{= NUll by default: initial value}

\item{xtol}{= 1.0e-5 by default: numerical accuracy}
}
\value{
\eqn{\Omega}
}
\description{
Minimize distance function to estimation \eqn{\hat{\Omega}}
}
