% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sym_mat}
\alias{sym_mat}
\title{Convert a matrix to a symmetric matrix}
\usage{
sym_mat(A)
}
\arguments{
\item{A}{a square matrix}
}
\value{
`A_sym`
}
\description{
Assign the upper triangular matrix to the lower triangle
}
