% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdpg_est.R
\name{estimate_ASE}
\alias{estimate_ASE}
\title{Function for estimation of P with ASE period by period}
\usage{
estimate_ASE(A_t, d, sim = NULL)
}
\arguments{
\item{A_t}{(n-by-n) Adjacency matrix}

\item{d}{dimension of the latent position}

\item{sim}{The random seed}
}
\value{
estimated P
}
\description{
Function for estimation of P with ASE period by period
}
\section{Warning}{
 random seeds changed
}

